重新组织项目结构，实现文件操作
